requirements.txt
numpy
pillow
opencv-python
tqdm
torch
torchvision
